uploaded homemade fim
In this script, a directory (not recursive) is scanned and each file is hashed. Hashes are written to a file. Each time run, the previous values are compared with the new values and if a change has been made to a file the file is noted as having been changed. Tested integrity with .csv, .txt, .ps1, and .py files.